Split identity pages
