+++ identity/webapp/pages/success.tsx
@@ -1,31 +1,24 @@
-import { GetServerSideProps, GetServerSidePropsContext, NextPage } from 'next';
+import { GetServerSidePropsContext, NextPage } from 'next';
 
 import { getServerData } from '@weco/common/server-data';
-import { SimplifiedServerData } from '@weco/common/server-data/types';
-import { AppErrorProps } from '@weco/common/services/app';
 import { serialiseProps } from '@weco/common/utils/json';
 import {
-  ContaineredLayout,
-  gridSize10,
-  gridSize8,
-} from '@weco/common/views/components/Layout';
-import Space from '@weco/common/views/components/styled/Space';
-import { usePageTitle } from '@weco/identity/hooks/usePageTitle';
-import { ApplicationReceived } from '@weco/identity/utils/copy';
-import PageWrapper from '@weco/identity/views/components/PageWrapper';
-import {
-  Container,
-  Wrapper,
-} from '@weco/identity/views/components/styled/layouts';
+  ServerSideProps,
+  ServerSidePropsOrAppError,
+} from '@weco/common/views/pages/_app';
+import SuccessPage, {
+  Props as SuccessPageProps,
+} from '@weco/identity/views/pages/success';
 
-type Props = {
-  serverData: SimplifiedServerData;
-  email: string;
+const Page: NextPage<SuccessPageProps> = props => {
+  return <SuccessPage {...props} />;
 };
 
-export const getServerSideProps: GetServerSideProps<
-  Props | AppErrorProps
-> = async (context: GetServerSidePropsContext) => {
+type Props = ServerSideProps<SuccessPageProps>;
+
+export const getServerSideProps: ServerSidePropsOrAppError<Props> = async (
+  context: GetServerSidePropsContext
+) => {
   const serverData = await getServerData(context);
 
   const { email } = context.query;
@@ -38,26 +31,4 @@ export const getServerSideProps: GetServerSideProps<
   };
 };
 
-const SuccessPage: NextPage<Props> = ({ email }) => {
-  usePageTitle('Application received');
-
-  return (
-    <PageWrapper title="Registration">
-      <ContaineredLayout gridSizes={gridSize10()}>
-        <Space $v={{ size: 'xl', properties: ['margin-top'] }}>
-          <Container>
-            <Wrapper>
-              <ContaineredLayout gridSizes={gridSize8()}>
-                <Space $v={{ size: 'xl', properties: ['padding-top'] }}>
-                  <ApplicationReceived email={email} />
-                </Space>
-              </ContaineredLayout>
-            </Wrapper>
-          </Container>
-        </Space>
-      </ContaineredLayout>
-    </PageWrapper>
-  );
-};
-
-export default SuccessPage;
+export default Page;
